# PERF: faster pd.concat when same concat float dtype but misaligned axis